requirements.txt
# Python 3.10 or higher

torch==2.7.0
torchvision==0.22.0
Pillow==11.2.1
gradio==5.29.0